requirements.txt
visen==0.0.1
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.10.0+cpu
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.11.0+cpu
-f https://download.pytorch.org/whl/torch_stable.html
paddlepaddle==2.4.2
shapely==1.7.0
pyclipper==1.2.0
fuzzywuzzy[speedup]==0.18.0
python-Levenshtein==0.12.2
unidecode==1.2.0
pdfminer.six==20201018
PyYAML==5.4.1
Pillow==9.3.0
opencv-python==3.4.9.33
pdf2image==1.6.0
PyPDF2==1.26.0
matplotlib==3.7.5
scikit-image==0.16.2
scipy==1.3.3
